+++ src/main/java/net/ow/movie/tmdb/feign/TMDBFeignClient.java
@@ -7,8 +7,8 @@ import net.ow.movie.tmdb.model.discover.TMDBDiscoverMovieRequest;
 import net.ow.movie.tmdb.model.discover.TMDBDiscoverTVShowRequest;
 import net.ow.movie.tmdb.model.genre.TMDBGenreList;
 import net.ow.movie.tmdb.model.movie.TMDBBaseMovie;
-import net.ow.movie.tmdb.model.movie.TMDBCredits;
 import net.ow.movie.tmdb.model.movie.TMDBMovie;
+import net.ow.movie.tmdb.model.person.TMDBCredits;
 import net.ow.movie.tmdb.model.search.TMDBSearchResult;
 import net.ow.movie.tmdb.model.trending.TMDBTrending;
 import net.ow.movie.tmdb.model.trending.TMDBTrendingMovie;
