+++ src/main/java/net/ow/movie/tmdb/model/tv/TMDBTVSeason.java
@@ -12,8 +12,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import net.ow.movie.tmdb.deserializer.ImagePathDeserializer;
-import net.ow.movie.tmdb.model.movie.TMDBCredits;
 import net.ow.movie.tmdb.model.person.TMDBCast;
+import net.ow.movie.tmdb.model.person.TMDBCredits;
 import net.ow.movie.tmdb.serializer.ImagePathSerializer;
 import net.ow.shared.jsonutils.deserializer.DateInstantDeserializer;
 import net.ow.shared.jsonutils.serializer.DateInstantSerializer;
