+++ src/main/java/net/ow/movie/tmdb/model/tv/TMDBTVShow.java
@@ -17,10 +17,10 @@ import net.ow.movie.tmdb.model.company.TMDBBaseCompany;
 import net.ow.movie.tmdb.model.country.TMDBCountry;
 import net.ow.movie.tmdb.model.genre.TMDBGenre;
 import net.ow.movie.tmdb.model.language.TMDBLanguage;
-import net.ow.movie.tmdb.model.movie.TMDBCredits;
 import net.ow.movie.tmdb.model.network.TMDBNetwork;
 import net.ow.movie.tmdb.model.person.TMDBCast;
 import net.ow.movie.tmdb.model.person.TMDBCreator;
+import net.ow.movie.tmdb.model.person.TMDBCredits;
 import net.ow.movie.tmdb.serializer.ImagePathSerializer;
 import net.ow.shared.jsonutils.deserializer.DateInstantDeserializer;
 import net.ow.shared.jsonutils.serializer.DateInstantSerializer;
